+++ human_design/__init__.py
@@ -3,6 +3,8 @@
 from .bodygraph import render_bodygraph_svg, write_bodygraph_svg
 from .career import generate_career_report, render_career_report_markdown
 from .empirical import analyze_forced_choice_experiment
+from .empirical_dataset import build_public_figure_manifest, load_manifest
+from .prediction_registry import analyze_prospective_registry
 from .engine import calculate_chart, parse_birth_datetime
 from .input import normalize_birth_input, normalize_birth_time_range
 from .product import build_llm_product
@@ -24,6 +26,9 @@ __all__ = [
     "build_llm_product",
     "build_timing_product",
     "analyze_forced_choice_experiment",
+    "analyze_prospective_registry",
+    "build_public_figure_manifest",
+    "load_manifest",
     "calculate_chart",
     "generate_career_report",
     "render_bodygraph_svg",